+++ Engine/Renderer/Vulkan/src/VKVertexBuffer.cpp
@@ -46,7 +46,7 @@ bool CVKVertexBuffer::BufferData(size_t offset, size_t size, const void* data)
 		return m_ptrBuffer->BufferData(offset, size, data);
 	}
 	else {
-		return m_ptrBuffer->BufferData((size_t)VKRenderer()->GetSwapChain()->GetFrameIndex() * m_size + offset, size, data);
+		return m_ptrBuffer->BufferData(VKRenderer()->GetSwapChain()->GetFrameIndex() * m_size + offset, size, data);
 	}
 }
